+++ R/mod_summarisedive.R
@@ -1,29 +1,57 @@
-#' summarisedive UI Function
+#' summarisedive server Function
 #'
-#' @description A shiny Module.
+#' @description A shiny Module to summarise dive objects.
+#' submodule of mod_01_squareUI
 #'
-#' @param id,input,output,session Internal parameters for {shiny}.
+#' @param id shiny id
+#' @param i18n traduction language in a reactive element
+#' @param dive a dive object created by the mn90 package
+#'
 #'
 #' @noRd 
-#'
-#' @importFrom shiny NS tagList 
-mod_summarisedive_ui <- function(id){
-  ns <- NS(id)
-  tagList(
- 
+mod_summarisediveServer <- function(id, i18n, dive){
+  moduleServer(
+    id,
+    ## Below is the module function
+    function(input, output, session) {
+      sup <- dive$palier$time > 0
+      n <- sum(sup)
+      diz <- sum(dive$palier$time[sup] > 9)
+      sp <- c(rep("  ", n - diz), rep(" ", diz))
+      
+      if(dive$palier$group != "Z"){
+        group <- paste0(i18n()$t('The dive group is '), dive$palier$group)
+      } else {
+        group <- ''
+      }
+      
+      if(dive$maj > 0){
+        maj <- paste0(i18n()$t('The dive majoration is '), dive$maj, ' minutes.')
+      } else {
+        maj <- ''
+      }
+      
+      ret <- renderText({
+        paste0(
+          i18n()$t("The first dive reach "), depth(dive),
+          i18n()$t(" meters for a duration of "), dtime(dive), " minutes.\n",
+          i18n()$t("Total dive time is "), round(diff(dive$hour),2),
+          i18n()$t(" minutes with an ascent of "), round(dive$dtr, 2), " minutes.\n",
+          sum(dive$palier$time > 0), i18n()$t(" stops :"),
+          paste(sprintf(
+            i18n()$t("%s%d minutes at %d meters"), sp,
+            dive$palier$time[sup],
+            dive$palier$depth[sup]
+          ),
+          collapse = "\n         "
+          ), '\n', group, '\n', maj
+        )
+        
+      })
+      ret
+    }
   )
 }
-    
-#' summarisedive Server Function
-#'
-#' @noRd 
-mod_summarisedive_server <- function(input, output, session){
-  ns <- session$ns
- 
-}
-    
-## To be copied in the UI
-# mod_summarisedive_ui("summarisedive_ui_1")
     
 ## To be copied in the server
 # callModule(mod_summarisedive_server, "summarisedive_ui_1")